port Model2D to new multi-d framework, kill Fitter2D
