+++ ionics_fits/__init__.py
@@ -6,5 +6,3 @@ from .normal import NormalFitter
 from . import models
 from . import utils
 from . import validators
-
-from . import multi_x
