--- ionics_fits/multi_x/__init__.py
@@ -1,2 +1,0 @@
-from .common import Model2D
-from .models import Cone2D, Gaussian2D, Parabola2D
